+++ CHANGELOG.md
@@ -2,6 +2,7 @@ UNRELEASED
 ----------
 
 * [ [#1005](https://github.com/digitalfabrik/integreat-cms/issues/1005) ] Add possibility to filter for unused media files
+* [ [#2048](https://github.com/digitalfabrik/integreat-cms/issues/2048) ] Add possibility to show media file usages in sidebar
 
 
 2023.3.0